Better but still with some nasty bugs
Code designed to pull code towards a CNT of 12000 pulls it towards 0.
Seems to be a long time before frame flips (30 seconds??), but they're
happening.  The time spend there may be less periodic.  Not sure,
haven't tested.  Going to try and strap to 12000 instead.  That's what
should be best, in theory.